Edit problems

+++ NET.S.2018.Lyubchenko.03/RootCalculator.Tests/RootCalculatetests.cs
@@ -10,7 +10,7 @@ namespace RootCalculaor.Tests
         public void FindNthRoot_1_5_00001_1()
         {
             double number = 1;
-            double degree = 5;
+            int degree = 5;
             double precision = 0.0001;
             double expected = 1;
 
@@ -23,7 +23,7 @@ namespace RootCalculaor.Tests
         public void FindNthRoot_8_3_00001_2()
         {
             double number = 8;
-            double degree = 3;
+            int degree = 3;
             double precision = 0.0001;
             double expected = 2;
 
@@ -36,7 +36,7 @@ namespace RootCalculaor.Tests
         public void FindNthRoot_0001_3_00001_2()
         {
             double number = 0.001;
-            double degree = 3;
+            int degree = 3;
             double precision = 0.0001;
             double expected = 0.1;
 
@@ -49,7 +49,7 @@ namespace RootCalculaor.Tests
         public void FindNthRoot_004100625_4_00001_045()
         {
             double number = 0.04100625;
-            double degree = 4;
+            int degree = 4;
             double precision = 0.0001;
             double expected = 0.45;
 
@@ -62,7 +62,7 @@ namespace RootCalculaor.Tests
         public void FindNthRoot_27_3_00001_3()
         {
             double number = 27;
-            double degree = 3;
+            int degree = 3;
             double precision = 0.0001;
             double expected = 3;
 
@@ -88,7 +88,7 @@ namespace RootCalculaor.Tests
         public void FindNthRoot_0004241979_9_000000001_0545()
         {
             double number = 0.004241979;
-            double degree = 9;
+            int degree = 9;
             double precision = 0.00000001;
             double expected = 0.545;
